can make page for all results, not able to yet make dynamic pages for events

+++ db/define_schema.rb
@@ -33,4 +33,5 @@ DB.define_column("signups", "competitor_id", "integer")
 
 DB.define_table("results")
 DB.define_column("results", "signup_id", "integer")
-DB.define_column("results", "time", "integer")
+DB.define_column("results", "time", "float")
+
